gcp/viet-soccer/youtube-upload/requirements.txt
google-api-python-client==2.5.0
google-cloud-pubsub==2.5.0
Flask==2.0.0
gunicorn==20.1.0
google-auth-oauthlib==0.4.4
